Movement and attacking fixed

+++ core/src/com/group25/game/EnemySpawner.java
@@ -67,7 +67,7 @@ public class EnemySpawner extends GameEntity {
      * @param img       - image of how the monster will look like
      * @param speed     - how quickly the monster travels
      */
-    public void spawnNewMonster(Level level, ArrayList<Enemy> enemies, int positionX, int positionY, int width,int height,int health, Sprite img,int speed){
+    public void spawnNewMonster(Level level, ArrayList<Enemy> enemies, int positionX, int positionY, int width,int height,int health, Sprite img,float speed){
         if(enemies.size()<spawnLimit){
             if(repeatedFrames == spawnTime){
                 repeatedFrames = 0;
